requirements.txt.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0

web3==6.11.4
eth-account==0.9.0
py-solc-x==1.1.1

cryptography==41.0.3
pycryptodome==3.19.0

requests==2.31.0

